### 2020-01-25 #2039 @whedon
[PRE REVIEW]: RADWave: Python code for ocean surface wave analysis by satellite radar altimeter
**Submitting author:** @CourtneySmith32 (<a href="http://orcid.org/0000-0002-9556-6626">Courtney Smith</a>)
**Repository:** <a href="https://github.com/pyReef-model/RADWave" target ="_blank">https://github.com/pyReef-model/RADWave</a>
**Version:** 0.02
