### 2020-01-25 #2039 @labarba
👋  @sjpfenninger — Could you handle this JOSS submission as editor?
(The subject area fits more closely with @kthyng, but she is editing 4 submissions right now.)
Have a look, and let me know.
