### 2020-03-06 pyReef-model/RADWave#1 @bradyrx
Tests break with pip install on python3.
I installed `RADWave` with pip as a python3 installation. My environment has all the necessary dependencies. When I run `python testInstall.py` I get the following error:

```python-traceback
Traceback (most recent call last):
  File "testInstall.py", line 4, in <module>
    stime=[1985,1,1], etime=[2018,12,31], test=True)
  File "/Users/ribr5703/miniconda3/envs/analysis/lib/python3.7/site-packages/RADWave/altiwave.py", line 156, in __init__
    ncs = NetCDFFile(satFile[0])
  File "netCDF4/_netCDF4.pyx", line 2291, in netCDF4._netCDF4.Dataset.__init__
  File "netCDF4/_netCDF4.pyx", line 1855, in netCDF4._netCDF4._ensure_nc_success
OSError: [Errno -45] NetCDF: Not a valid data type or _FillValue type mismatch: b'http://thredds.aodn.org.au/thredds/dodsC/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/040S_140E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_036S-154E-DM00.nc\n'
```

It looks like it's somehow bringing the new line character `\n` into the input.
