### 2020-03-06 #2083 @bradyrx
# Review of `RADWave`

Thank you to @CourtneySmith32 and colleagues for submitting this software and manuscript. The documentation is extensive and I found the binder examples impressive and quite useful. I appreciate that all of the binder analysis was done with THREDDS to show that data can be pulled from remote servers and need not be stored locally for this package to work. It's clear that `RADWave` serves a need in the community and will make analyzing altimetry data a lot smoother. Nice work!

I have comments below on the software, documentation, and summary paper that should be addressed prior to publication with JOSS. Generally, some more care needs to be taken with package management (testing, CI, etc.) and the code base needs to be cleaned up a bit (consistent naming and returns, linting, etc.). However, once these are addressed I think it's a great fit for the journal.

## General Comments/Questions

* I would consider renaming the package to `radwave`. PEP guidelines suggest that python packages should be named in all lowercase, and only use underscores to improve readability (https://www.python.org/dev/peps/pep-0008/#package-and-module-names).
* Does this only work with the setup from AODN, or will it work with any THREDDS altimetry data? From the summary paper, it seems like AODN hosts all altimetry mission output, so this might not be relevant. If that's the case, I would suggest making this clear in the docs. Currently the docs/manuscript say AODN is used to illustrate the package capabilities, but has any testing been done on other servers?
* Why are the notebooks/tutorial replicated in both `/Notebooks/` and `/src/RADWave/Notebooks/`? I imagine these can be consolidated into one location. Typically for python, everything from `src` can be moved to the main folder.
* I would encourage the use of a CI service like Travis for this package. This ensures that the package can be built/installed for a given PR's updates, and will run a testing suite as well (which I suggest in the next section).
* Is 30 days the common moving average window in the field for wave time series? Currently, the 30-day average is hard-coded into the class methods for plotting time series. It might be useful for the user to be able to designate the moving average window length.

## Code Comments

* I don't think the empty `return` statements are necessary throughout the class methods.
* I would strongly recommend running `flake8` or a similar linter on the code. There's currently 100 errors that come through from `flake8` on the code base. It might also help to have a column-width limitation for user and developer readability. E.g. `black`, or just enforce an 80 or 88 line limit.
* I think functions can generally be renamed for clarity. E.g. `processingAltimeterData` can become `processAltimeterData` and `readingAltimeterData` to `readAltimeterData`. `seriesSeasonMonth` can be renamed something more clear like `computeSeasonalCharacteristics` (or something less verbose). `timeSeries` should be something more descriptive like `generate_time_series` or something in camelCase. It currently reads like a class attribute.
* Time series options (`'H', 'T', 'P', 'E' and 'Cg' `) should be clearly defined in the docs and/or in the docstring for `plotTimeSeries`.
* I would consider returning the derived `df`'s from functions instead of just adding them as object attributes. E.g. from `timeSeries` and `close2Track`. Generally it's intuitive for the user to be returned the computation result from a method like these. Since there's no `__repr__` on the main class, the user has no interaction with their object and can't see when `timeSeries` has already been computed. Further, there's inconsistencies with `seriesSeasonMonth` which returns the `df`, but the former two mentioned methods which don't.
* `cycloneAltiPoint` should probably be renamed to `plotCycloneAltiPoint` for consistency and clarity.

## Review Checklist

### General checks

Could the lead author please outline their contribution and Dr. Vila-Concejo's contribution to the software? The repository suggests that all of the code was written by @tristan-salles, and the Sphinx copyright and contact email on the documentation goes to @tristan-salles as well.

### Documentation

    I think the installation page is comprehensive and I appreciate the multiple methods to install. However, I would also suggest registering `RADWave` with `conda-forge`, since this is probably the best package management solution out there right now.

    The only testing available is `testInstall.py`, which seems to just test that `RADWave` can be imported and the main THREDDS method run. First, it's currently breaking for me (see https://github.com/pyReef-model/RADWave/issues/1).

    Secondly, `RADWave` should have unit testing well beyond this. I would suggest having, e.g. a `pytest` suite that is run with continuous integration, such as Travis, on all PRs. You can use `coveralls` or a similar service to check how much of the code base is covered by testing. At the minimum, I would test all of the methods in the `waveAnalysis` object and ensure that they run without breaking. You should probably use `pytest.parametrize` to loop through different keyword options as well.

    If the authors don't have too much experience with this level of unit testing, I suggest checking out our testing suite at `climpred`: https://github.com/bradyrx/climpred/tree/master/climpred/tests. I think we have generally minimal examples there, which are easier to base testing off of than a bigger package like `numpy` or `xarray`.

### Software paper

    Please see comments below. I suggest that the paper be carefully revised in a few locations for grammar and clarity.

    Please see comments below. I suggest the addition of at least one reference in the summary.

#### Summary
* In the first sentence, I would restructure it to either say that satellite radar **altimeters** are excellent instruments or that satellite radar **altimetry** is an excellent remote sensing technique.
* In the second sentence, gramatically it should be restructured to say "These measurements can be used in conjunction with one another ..."
* As in the docs, I would suggest citing Chelton 2001 or something similar to back up the gap in satellite operation.
* Please rewrite the following sentence for clarity:

> Consequently, analysis of this long-term temporal record can lead to significant insights into inter-annual, seasonal and decadal variations in wave yearly seasonality and decadal trend.

I'm not exactly sure what is being said here. E.g., what is "wave yearly seasonality"? How do decadal trends fit into this picture? Perhaps that analyzing the record can lead to insights in interannual, seasonal, and decadal variations in wave statistics (or something similar).

#### Main Text

* Please reformulate this sentence as well. It seems like there are words missing (e.g. "allows the user to query"?):

> RADWave allows to query over a range of spatial and temporal scales altimeter parameters in specific geographical regions and subsequently calculates significant wave heights, periods, group velocities, average wave energy densities and wave energy fluxes.

* Perhaps it would be good to follow the line about there being no open-source code available for this purpose with a suggestion that users currently have to download the data manually (or write by hand THREDDS retrieval code) and write custom analysis and visualization functions using `pandas`, `numpy`, `matplotlib`, etc. And that `RADWave` combines this all into a simple object-oriented framework.
