### 2020-03-09 24b3599
update based on JOSS review
