+++ src/RADWave/altiwave.py
@@ -121,11 +121,9 @@ class waveAnalysis(object):
         bbox=None,
         stime=None,
         etime=None,
-        satNames=None,
-        test=False,
+        satNames=None
     ):
 
-        self.test = test
         if cycloneCSV is not None:
             try:
                 with open(str(cycloneCSV)) as file:
@@ -210,18 +208,6 @@ class waveAnalysis(object):
                                  after the end time 'etime'."
                 )
 
-    def _test_progress(self, job_title, progress):
-
-        length = 20
-        block = int(round(length * progress))
-        msg = "\r{0}: [{1}] {2}%".format(
-            job_title, "#" * block + "-" * (length - block), round(progress * 100, 2)
-        )
-        if progress >= 1:
-            msg += " DONE\r\n"
-        sys.stdout.write(msg)
-        sys.stdout.flush()
-
     def _extractURLsatellite(self, fileURL, satName):
         """
         Function to convert the list of URL’s generated by the IMOS portal to a
@@ -296,8 +282,7 @@ class waveAnalysis(object):
             saveCSV (str): filename used to save processed altimeter data obtained from the OPeNDAP web service [default: 'altimeterData.csv']
         """
 
-        if not self.test:
-            print("Processing Altimeter Dataset \n")
+        print("Processing Altimeter Dataset \n")
 
         t0 = time.clock()
         p = 0
@@ -311,12 +296,8 @@ class waveAnalysis(object):
             if altimeter_pick == self.nameSat[u] or altimeter_pick == "all":
                 for k in range(len(picked_url)):
 
-                    if self.test:
-                        step += 1
-                        self._test_progress("Test RADWave installation:", step / steps)
-
                     ncs = NetCDFFile(picked_url[k])
-                    if k == 0 and not self.test:
+                    if k == 0:
                         print(
                             "   +  name {:<11s} / number of tracks \
                               {:<4d}".format(
@@ -421,41 +402,35 @@ class waveAnalysis(object):
                             p += 1
 
         if p > 0:
-            if not self.test:
-                self.saveCSV = saveCSV
-                combineframe.to_csv(
-                    str(self.saveCSV),
-                    columns=["lat", "lon", "wh", "time", "ws"],
-                    sep=" ",
-                    index=False,
-                    header=1,
-                )
-                altiData = pd.read_csv(
-                    str(self.saveCSV),
-                    sep=r"\s+",
-                    engine="c",
-                    header=0,
-                    na_filter=False,
-                    low_memory=False,
-                )
-                data = altiData.sort_values(by=["time"])
-                data = data.replace(r"^\s*$", np.nan, regex=True)
-                data = data.dropna()
-                self.lat = np.asarray(data.values[:, 0], dtype=np.float64)
-                self.lon = np.asarray(data.values[:, 1], dtype=np.float64)
-                self.wh = np.asarray(data.values[:, 2], dtype=np.float64)
-                self.times = data.values[:, 3]
-                self.ws = np.asarray(data.values[:, 4], dtype=np.float64)
-                print(
-                    " \nProcessing altimeter dataset took: ",
-                    int(time.clock() - t0),
-                    "s",
-                )
-            else:
-                print(
-                    "\nAll tests passed - RADWave installation is \
-                      completed !\n"
-                )
+            self.saveCSV = saveCSV
+            combineframe.to_csv(
+                str(self.saveCSV),
+                columns=["lat", "lon", "wh", "time", "ws"],
+                sep=" ",
+                index=False,
+                header=1,
+            )
+            altiData = pd.read_csv(
+                str(self.saveCSV),
+                sep=r"\s+",
+                engine="c",
+                header=0,
[…]
