+++ src/requirements.txt
@@ -1,4 +1,5 @@
 numpy>=1.15.0
+pytest
 six>=1.11.0
 setuptools>=38.4.0
 pandas>=0.25
