+++ src/.travis.yml
@@ -1,0 +1,9 @@
+
+language: python
+
+python:
+ - "3.5"
+
+install: pip install -e .
+
+script: pytest

[…]
