### 2020-03-10 0e9361e
update travis pip install
