+++ .travis.yml
@@ -4,6 +4,6 @@ language: python
 python:
  - "3.5"
 
-install: pip install -e .
+install: pip install -e src/
 
 script: pytest
