+++ .travis.yml
@@ -17,8 +17,7 @@ install:
   - "python -m pip install coveralls>=1.1"
   - "pip install -e src/"
 
-script:
-  - python -m pytest -v -m --cov rasterio --cov-report term-missing
-
+script: pytest
+  
 after_success:
   - coveralls || echo "!! intermittent coveralls failure"
