+++ .travis.yml
@@ -13,11 +13,12 @@ addons:
       - libgdal-dev
 
 install:
+  - "python -m pip install --upgrade numpy"
   - "python -m pip install pyproj"
   - "python -m pip install coveralls>=1.1"
   - "pip install -e src/"
 
 script: pytest
-  
+
 after_success:
-  - coveralls || echo "!! intermittent coveralls failure"
+  - coveralls
