+++ .travis.yml
@@ -15,10 +15,12 @@ addons:
 install:
   - "python -m pip install --upgrade numpy"
   - "python -m pip install pyproj"
+  - "python -m pip install pytest-cov"
   - "python -m pip install coveralls>=1.1"
   - "pip install -e src/"
 
-script: pytest
+script:
+  -- pytest --cov=src/
 
 after_success:
   - coveralls
