### 2020-03-10 pyReef-model/RADWave#1 @tristan-salles
This has been fixed we now use pytest and Travis CI for our testing
