+++ joss/paper.bib
@@ -1,5 +1,0 @@
-@article{Ribal201933Observations, title = {{33 Years of Globally Calibrated Wave Height and Wind Speed Data Based on Altimeter Observations}}, year = {2019}, author = {Ribal, Agustinus and Young, Ian R}, number = {January} }
-
-@article{Young2018OnObservations, title = {{On the determination of global ocean wind and wave climate from satellite observations}}, year = {2018}, journal = {Remote Sensing of Environment}, author = {Young, I.R. and Donelan, M.A.}, month = {9}, pages = {228--241}, volume = {215}, publisher = {Elsevier}, doi = {10.1016/J.RSE.2018.06.006} }
-
-@misc{author = {{The Australian Ocean Data Network}}, title = {IMOS - SRS Surface Waves Sub-Facility - altimeter wave/wind}, year = {2019}, note = {url{https://portal.aodn.org.au/search}}}
