+++ src/RADWave/altiwave.py
@@ -3,18 +3,12 @@
 # Copyright (c) 2020 Tristan Salles
 # Licensed under the GNU LGPL Version 3
 
-import io
-import gc
 import re
-import sys
 import math
 import time
-import errno
-import requests
 import numpy as np
 import pandas as pd
 
-import scipy
 from functools import reduce
 from scipy.spatial import cKDTree as _cKDTree
 
@@ -22,10 +16,9 @@ from scipy.spatial import cKDTree as _cKDTree
 try:
     import netCDF4
     from netCDF4 import Dataset as NetCDFFile
-    from netCDF4 import num2date, date2num, date2index
     import datetime as dt
 except ImportError:
-    print('netCDF4 is required and needs to be installed via pip')
+    print("netCDF4 is required and needs to be installed via pip")
     pass
 
 # For readthedoc...
@@ -36,21 +29,21 @@ try:
     import cartopy.io.img_tiles as cimgt
     from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
 except ImportError:
-    print('cartopy is required and needs to be installed via pip')
+    print("cartopy is required and needs to be installed via pip")
     pass
 
 # For readthedoc...
 try:
     import pymannkendall as mk
 except ImportError:
-    print('pymannkendall is required and needs to be installed via pip')
+    print("pymannkendall is required and needs to be installed via pip")
     pass
 
 # For readthedoc...
 try:
     import seaborn as sns
 except ImportError:
-    print('seaborn is required and needs to be installed via pip')
+    print("seaborn is required and needs to be installed via pip")
     pass
 
 # For readthedoc...
@@ -58,44 +51,51 @@ try:
     import shapely
     from shapely.geometry.polygon import LinearRing
 except ImportError:
-    print('shapely is required and needs to be installed via pip')
+    print("shapely is required and needs to be installed via pip")
     pass
 
 # For readthedoc...
 try:
-     import geopy.distance
+    import geopy.distance
 except ImportError:
-    print('geopy is required and needs to be installed via pip')
+    print("geopy is required and needs to be installed via pip")
     pass
 
 import matplotlib.pyplot as plt
-import matplotlib.ticker as mticker
 from matplotlib.transforms import offset_copy
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 from pandas.plotting import register_matplotlib_converters
+
 register_matplotlib_converters()
 
-plt.rcParams['mathtext.fontset'] = 'cm'
+plt.rcParams["mathtext.fontset"] = "cm"
+
 
 class waveAnalysis(object):
     """
-    **RadWave** is a package to perform wave analysis using *Radar Altimeter Data*. Radar altimeters are mounted on satellites
-    and measure a footprint of the ocean directly under the satellite (approx. 5-7 km wide). When the water surface is calm and flat,
-    the reflection of the radar pulse back to the altimeter is almost instantaneous. In contrast, when waves are present the pulse
-    is first reflected at the crest of the wave, then progressively reflected as the pulse reaches the wave trough.
+    **RadWave** is a package to perform wave analysis using AODN *Radar
+    Altimeter Data*. Radar altimeters are mounted on satellites and measure a
+    footprint of the ocean directly under the satellite (approx. 5-7 km wide).
+    When the water surface is calm and flat, the reflection of the radar pulse
+    back to the altimeter is almost instantaneous. In contrast, when waves are
+    present the pulse is first reflected at the crest of the wave, then
+    progressively reflected as the pulse reaches the wave trough.
 
     Method:
-        Calibrated altimeter dataset provide **wave height** and **wind speed** through a series of equations described in
-        [RibalYoung2019]
[…]
