+++ Docker/Dockerfile
@@ -15,7 +15,9 @@ RUN apt-get update -qq && \
     gettext && \
     apt-get clean && rm -rf /var/lib/apt/lists/*
 
-### bioLEC - Notebooks
+### RADWave - Notebooks
+RUN python3 -m pip install --no-cache-dir \
+    radwave==0.0.7
 
 ENV MODULE_DIR="src"
 ADD --chown=jovyan:jovyan $MODULE_DIR $MODULE_DIR

[…]
