+++ Notebooks/dataset/IMOS_YASI_east.txt
@@ -1,94 +1,0 @@
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/CRYOSAT-2/020S_160E/IMOS_SRS-Surface-Waves_MW_CRYOSAT-2_FV02_013S-172E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_013S-173E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_013S-171E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/CRYOSAT-2/020S_160E/IMOS_SRS-Surface-Waves_MW_CRYOSAT-2_FV02_014S-170E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/ENVISAT/020S_160E/IMOS_SRS-Surface-Waves_MW_ENVISAT_FV02_013S-173E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-1/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-1_FV02_016S-174E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/CRYOSAT-2/020S_160E/IMOS_SRS-Surface-Waves_MW_CRYOSAT-2_FV02_016S-173E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_016S-170E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_017S-172E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_013S-170E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-1/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-1_FV02_013S-170E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-1/020S_160E/IMOS_SRS-Surface-Waves_MW_JASON-1_FV02_013S-174E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/ENVISAT/020S_160E/IMOS_SRS-Surface-Waves_MW_ENVISAT_FV02_015S-171E-DM00.nc
[…]
