+++ Notebooks/dataset/IMOS_YASI_west.txt
@@ -1,76 +1,0 @@
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/ENVISAT/020S_140E/IMOS_SRS-Surface-Waves_MW_ENVISAT_FV02_014S-157E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_015S-157E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_015S-159E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_015S-156E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/CRYOSAT-2/020S_140E/IMOS_SRS-Surface-Waves_MW_CRYOSAT-2_FV02_016S-156E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_016S-158E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/CRYOSAT-2/020S_140E/IMOS_SRS-Surface-Waves_MW_CRYOSAT-2_FV02_015S-156E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-1/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-1_FV02_016S-158E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_013S-159E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/ENVISAT/020S_140E/IMOS_SRS-Surface-Waves_MW_ENVISAT_FV02_013S-158E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-1/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-1_FV02_015S-158E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/JASON-2/020S_140E/IMOS_SRS-Surface-Waves_MW_JASON-2_FV02_014S-159E-DM00.nc
-http://data.aodn.org.au/IMOS/SRS/Surface-Waves/Wave-Wind-Altimetry-DM00/ENVISAT/020S_140E/IMOS_SRS-Surface-Waves_MW_ENVISAT_FV02_015S-156E-DM00.nc
[…]
