+++ Dockerfile
@@ -26,8 +26,8 @@ RUN ipython3 -c 'import RADWave; RADWave.documentation.install_documentation(pat
 ADD  --chown=jovyan:jovyan Notebooks/0-StartHere.ipynb Notebooks/0-StartHere.ipynb
 
 
-RUN python3 -m pip install --no-cache-dir --upgrade nose \
-   jupyterhub 
+RUN python3 -m pip install --no-cache-dir --upgrade nose geopy \
+   jupyterhub
 
 RUN mkdir -p /usr/local/files && chown -R jovyan:jovyan /usr/local/files
 ADD --chown=jovyan:jovyan Docker/scripts  /usr/local/files
