### 2020-03-10 #2083 @scivision
Review checklist for @scivision
Conflict of interest
[x] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
Code of Conduct
 I confirm that I read and will adhere to the JOSS code of conduct.
General checks
 Repository: Is the source code for this software available at the repository url?
 License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
 Contribution and authorship: Has the submitting author (@CourtneySmith32) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
Functionality
 Installation: Does installation proceed as outlined in the documentation?
 Functionality: Have the functional claims of the software been confirmed?
 Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
Documentation
 A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
 Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
 Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
 Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
 Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
Software paper
 Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
 A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
 State of the field: Do the authors describe how this software compares to other commonly-used packages?
 Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
 References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
