+++ joss/paper.bib
@@ -1,4 +1,13 @@
 
+@book{Chelton:2001,
+  author    = {Chelton, D. B., Ries, J. C., Haines, B. J., Fu, L. L. and Callahan, P. S.},
+  title     = {Satellite Altimetry, Satellite altimetry and Earth sciences},
+  editor    = {Fu L.-L. and Cazenave A.},
+  publisher = {International Geophysics Series, Academic Press, San Diego},
+  year      = 2001,
+  volume    = 69,
+  isbn      = {9780080516585}
+}
 
 @article{Young:2018,
    author = {Young, I. R. and Donelan, M.},
