### 2020-03-10 #2083 @tristan-salles
We would like to thank @bradyrx for his really useful and helpful comments. **They have all helped to improve significantly the quality of the package!**

Below is a point by point revisions of his suggestions.

**General Comments/Questions**

+ I would consider renaming the package to radwave. PEP guidelines suggest that python packages should be named in all lowercase, and only use underscores to improve readability (https://www.python.org/dev/peps/pep-0008/#package-and-module-names).

> we have changed the package name to radwave as suggested by the reviewer.

+ Why are the notebooks/tutorial replicated in both /Notebooks/ and /src/RADWave/Notebooks/? I imagine these can be consolidated into one location. Typically for python, everything from src can be moved to the main folder.

>  We have removed the duplicated tutorials.

+ I would encourage the use of a CI service like Travis for this package. This ensures that the package can be built/installed for a given PR's updates, and will run a testing suite as well (which I suggest in the next section).

> We are now using Travis with pytest for RADwave

+ Is 30 days the common moving average window in the field for wave time series? Currently, the 30-day average is hard-coded into the class methods for plotting time series. It might be useful for the user to be able to designate the moving average window length.

> Monthly (30 days) average window is commonly used for wave time series analysis saying that one might want to use another time interval and following reviewer’s comment we have modified the code to allow user to pick a specific number of days in the function timeSeries that now accepts an argument: def timeSeries(self, days=30)

+ Does this only work with the setup from AODN, or will it work with any THREDDS altimetry data? From the summary paper, it seems like AODN hosts all altimetry mission output, so this might not be relevant. If that's the case, I would suggest making this clear in the docs. Currently the docs/manuscript say AODN is used to illustrate the package capabilities, but has any testing been done on other servers?

> As pointed out by the reviewer, the package in its current form only works with AODN altimetry data. In a future version of the code we could make it work with other THREDDS server but as AODN currently hosts all altimetry missions this is not an issue for now. Following reviewer advice we have making it clear in the docs.

**Code Comments**

+ I don't think the empty return statements are necessary throughout the class methods.

> As suggested by the reviewer, we have removed the empty return statements.

+ I would strongly recommend running flake8 or a similar linter on the code. There's currently 100 errors that come through from flake8 on the code base. It might also help to have a column-width limitation for user and developer readability. E.g. black, or just enforce an 80 or 88 line limit.

> Following reviewer’s comment we have used flake8 and black to reformat the main code contained in altiwave.py. There are still some line running over the 80 line mark but they are mostly comments and documentation related parts of the files.

+ I think functions can generally be renamed for clarity. E.g. processingAltimeterData can become processAltimeterData and readingAltimeterData to readAltimeterData. seriesSeasonMonth can be renamed something more clear like computeSeasonalCharacteristics (or something less verbose). timeSeries should be something more descriptive like generate_time_series or something in camelCase. It currently reads like a class attribute.

> We have renamed all the functions based on the reviewer’s comment: e.g. processAltimeterData, readAltimeterData, computeSeasonalCharacteristics and generate_time_series.

+ Time series options ('H', 'T', 'P', 'E' and 'Cg' ) should be clearly defined in the docs and/or in the docstring for plotTimeSeries.

> We have added the options in the docs and the docstring for clarity.

+ I would consider returning the derived df's from functions instead of just adding them as object attributes. E.g. from timeSeries and close2Track. Generally it's intuitive for the user to be returned the computation result from a method like these. Since there's no __repr__ on the main class, the user has no interaction with their object and can't see when timeSeries has already been computed. Further, there's inconsistencies with seriesSeasonMonth which returns the df, but the former two mentioned methods which don't.

> Following reviewer’s comment we are now returning derived data frames for the generate_time_series and close2Track functions.

+ cycloneAltiPoint should probably be renamed to plotCycloneAltiPoint for consistency and clarity.

> The function has been renamed accordingly.

**Review Checklist**

**General checks**

+ Contribution and authorship: Has the submitting author (@CourtneySmith32) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? Could the lead author please outline their contribution and Dr. Vila-Concejo's contribution to the software? The repository suggests that all of the code was written by @tristan-salles, and the Sphinx copyright and contact email on the documentation goes to @tristan-salles as well.

> The submitting author CS designed an initial version of RADWave as part of her Honours under the supervision of AVC and TS. Subsequently TS wrapped the main functions and extended the code functionalities. The tests were designed by all the three co-authors.

**Documentation**

+ Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. I think the installation page is comprehensive and I appreciate the multiple methods to install. However, I would also suggest registering RADWave with conda-forge, since this is probably the best package management solution out there right now.

> The package is available via pypi and can easily be installed in Anaconda using the pip command. We have decided to leave it as is for now as this does not seem to add an extra installation steps to the user as far as we can see plus it limits the number of packages that we will have to maintain in the future.

+ Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? The only testing available is testInstall.py, which seems to just test that RADWave can be imported and the main THREDDS method run. First, it's currently breaking for me (see pyReef-model/RADWave#1).

> This has been fixed, we now use Travis and pytest for automated tests.

 + Secondly, RADWave should have unit testing well beyond this. I would suggest having, e.g. a pytest suite that is run with continuous integration, such as Travis, on all PRs. You can use coveralls or a similar service to check how much of the code base is covered by testing. At the minimum, I would test all of the methods in the waveAnalysis object and ensure that they run without breaking. You should probably use pytest.parametrize to loop through different keyword options as well. If the authors don't have too much experience with this level of unit testing, I suggest checking out our testing suite at climpred: https://github.com/bradyrx/climpred/tree/master/climpred/tests. I think we have generally minimal examples there, which are easier to base testing off of than a bigger package like numpy or xarray.

> Following reviewer’s comment we have added the pytest suite to our package and use Travis for continuous integration as well as coveralls.

**Software paper**

+ In the first sentence, I would restructure it to either say that satellite radar altimeters are excellent instruments or that satellite radar altimetry is an excellent remote sensing technique.

> We have followed reviewer’s comment and changed the first sentence accordingly.

+ In the second sentence, gramatically it should be restructured to say "These measurements can be used in conjunction with one another ..." As in the docs, I would suggest citing Chelton 2001 or something similar to back up the gap in satellite operation.

> We have modified the sentence and added the Chelton 2001 reference.

+ Consequently, analysis of this long-term temporal record can lead to significant insights into inter-annual, seasonal and decadal variations in wave yearly seasonality and decadal trend.

> The sentence has been modified accordingly.

+ Please reformulate this sentence as well. It seems like there are words missing (e.g. "allows the user to query"?): RADWave allows to query over a range of spatial and temporal scales altimeter parameters in specific geographical regions and subsequently calculates significant wave heights, periods, group velocities, average wave energy densities and wave energy fluxes.

> We have followed reviewer’s comment and changed these sentences accordingly in the paper and the documentation.

+ Perhaps it would be good to follow the line about there being no open-source code available for this purpose with a suggestion that users currently have to download the data manually (or write by hand THREDDS retrieval code) and write custom analysis and visualization functions using pandas, numpy, matplotlib, etc. And that RADWave combines this all into a simple object-oriented framework.

> We have added a sentence about this in the paper.
