### 2020-03-10 #2083 @bradyrx
Thank you for the quick turnaround @tristan-salles! I suggest that this is accepted/published after addressing the two minor comments below:

> We have renamed all the functions based on the reviewer’s comment: e.g. processAltimeterData, readAltimeterData, computeSeasonalCharacteristics and generate_time_series.

I would make it `generateTimeSeries` for consistency with camelCase. I should have noted that.

> While domain experts are able to extract satellite radar altimetry dataset, they currently have to download the data manually and write custom analysis and visualization functions using pandas, numpy, matplotlib to extract wave characteristics.

The above is from the updated text. There's still some grammatical issues here. I would change it to:

"While domain experts are able to assess satellite radar altimetry datasets, they currently have to download the data manually and write custom analysis and visualization functions using, e.g., `numpy`, `pandas`, and `matplotlib` to extract wave characteristics.

**And here's some ranting thoughts on package management below**. You don't need to address these comments, but I thought they might help moving forward.

> Following reviewer’s comment we have used flake8 and black to reformat the main code contained in altiwave.py. There are still some line running over the 80 line mark but they are mostly comments and documentation related parts of the files.

We use a `black` line limit of 88, and impose that on `black` and `flake8` in our setup.cfg file here: https://github.com/bradyrx/climpred/blob/master/setup.cfg. 80 ends up pretty short. Anyways, with whatever line limit you use, you can manually force your comments to adhere to this limit. For long comment lines just make them multi-lined. For docstring things (like arguments) if they are multiline, just tab out the second line like a hanging indent.

This all sounds really tedious but makes the code more readable for future developers. We also have a `run_linter.sh` script (https://github.com/bradyrx/climpred/blob/master/ci/run-linter.sh) that we run at the end of our Travis to make sure everything looks good for new code. Note that we add all of our code through Pull Requests so that one can't merge the new code until it is cleaned up like this.

Not that `climpred` is the bible for this. :) We've set up all of these CI/pre-commit workflows from advice from other packages. It's tedious to get going, but then makes the workflow very straight forward. `pre-commit` totally changed our development (https://github.com/bradyrx/climpred/blob/master/.pre-commit-config.yaml). It's worth trying out to see how awesome it is! Once you do `pip install pre-commit`, have a config file like I linked in your repo and do `pre-commit install`. Then you can set up a bunch of rules (like in my example) and every time you try to commit code it will run those checks. The `black`/`flake8` hook on there just runs it automatically as you commit the code so you don't even have to think about it.
