### 2020-03-10 db3b5a7
change generate_time_series to generateTimeSeries
