+++ joss/paper.md
@@ -36,7 +36,7 @@ Additionally, the high spatial resolution provided by altimeter data is particul
 
 **RADWave** is Python package that provides a mechanism to access altimeter datasets through web-enabled data services (THREDDS). The package capabilities are illustrated based on the the Australian Ocean Data Network global database [@Ribal:2019]. **RADWave** allows  the user to query over a range of spatial and temporal scales altimeter parameters in specific geographical regions and subsequently calculates significant wave heights, periods, group velocities, average wave energy densities and wave energy fluxes.
 
-While domain experts are able to extract satellite radar altimetry dataset, they currently have to download the data manually and write custom analysis and visualization functions using pandas, numpy, matplotlib to extract wave characteristics. **RADWave** combines this all into a simple object-oriented framework.
+While domain experts are able to assess satellite radar altimetry datasets, they currently have to download the data manually and write custom analysis and visualization functions using, e.g., numpy, pandas, and matplotlib to extract wave characteristics. **RADWave** combines this all into a simple object-oriented framework.
 
 **RADWave** can be used to easily calculate past wave conditions and infers long term wave climate variability, providing new insights on wave modal conditions, seasonal changes, long-term trends and associated modulation by climate oscillations. It can also be used to assess locally the impact of wave-generated cyclones in offshore areas.
 
