+++ src/docs/conf.py
@@ -39,7 +39,7 @@ author = u'Tristan Salles'
 # The short X.Y version
 version = u''
 # The full version, including alpha/beta/rc tags
-release = u'0.0.1'
+release = u'0.0.8'
 
 
 # -- General configuration ---------------------------------------------------
