+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "radwave" %}
-{% set version = "0.0.4" %}
+{% set version = "0.0.8" %}
 
 package:
   name: "{{ name|lower }}"
