### 2020-03-10 #2083 @tristan-salles
Thanks! The 2 comments have been resolved and the docker image regenerated and tested as well as the binder one! Again thanks for all the feedbacks and insights!
