+++ src/setup.py
@@ -50,5 +50,7 @@ if __name__ == "__main__":
                                     'Notebooks/images/*']},
           include_package_data=True,
           classifiers=['Programming Language :: Python :: 3.5',
-                       'Programming Language :: Python :: 3.6']
+                       'Programming Language :: Python :: 3.6',
+                       'Programming Language :: Python :: 3.7',
+                       'Programming Language :: Python :: 3.8']
           )
