### 2020-03-10 #2083 @tristan-salles
Thanks @scivision for the reviews & suggestions. We have addressed all your comments.

+ The package uses the src/ directory as the top level for Python code, which is a popular layout choice. However, I believe such package still generally keep the setup.cfg and setup.py at the topmost level of the repo, as the repo won't install without descending into the src/ directory, which would be remediated by moving setup.py and setup.cfg up to the top.

> We have moved the setup.py file to the topmost level of the repo.

+ The Classifier metadata goes down to Python 3.3, but even Python 3.4 is challenging to use for the past year or more. The CI is set to Python 3.5 and probably the classifier metadata could also be set from Python 3.5 to 3.8.

> We have changed the classifier to be set from Python 3.5 to 3.8

+ Automated test: please consider pytest.approx instead of using round() to compare

> We have changed round to pytest.approx as suggested by the reviewer and have checked that Travis CI is still building without issues.
 on your comments we
