### 2020-03-27 #2083 @sjpfenninger
@tristan-salles You seem to have duplicated setup.py and setup.cfg (it's at both the top level and inside src/) -- are you sure this is intended? It ought to only be in one place.
