### 2020-03-27 d7db1d5
remove duplicate setup configuration files
