+++ src/docs/conf.py
@@ -14,17 +14,21 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('../'))
-sys.path.insert(0, os.path.abspath('../src'))
+
+sys.path.insert(0, os.path.abspath("."))
+sys.path.insert(0, os.path.abspath("../"))
+sys.path.insert(0, os.path.abspath("../src"))
 
 import sphinx_rtd_theme
 
-from sphinx.builders.html import (StandaloneHTMLBuilder, DirectoryHTMLBuilder,
-                                  SingleFileHTMLBuilder)
+from sphinx.builders.html import (
+    StandaloneHTMLBuilder,
+    DirectoryHTMLBuilder,
+    SingleFileHTMLBuilder,
+)
 
 # Redefine supported_image_types for the HTML builder
-html_img_types = ['image/gif', 'image/svg+xml', 'image/png', 'image/jpeg']
+html_img_types = ["image/gif", "image/svg+xml", "image/png", "image/jpeg"]
 StandaloneHTMLBuilder.supported_image_types = html_img_types
 DirectoryHTMLBuilder.supported_image_types = html_img_types
 SingleFileHTMLBuilder.supported_image_types = html_img_types
@@ -32,14 +36,14 @@ SingleFileHTMLBuilder.supported_image_types = html_img_types
 
 # -- Project information -----------------------------------------------------
 
-project = u'RADWave'
-copyright = u'2020, Tristan Salles'
-author = u'Tristan Salles'
+project = u"RADWave"
+copyright = u"2020, Tristan Salles"
+author = u"Tristan Salles"
 
 # The short X.Y version
-version = u''
+version = u""
 # The full version, including alpha/beta/rc tags
-release = u'0.0.8'
+release = u"0.0.8"
 
 
 # -- General configuration ---------------------------------------------------
@@ -51,19 +55,23 @@ release = u'0.0.8'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.viewcode', 'sphinx.ext.todo',
-              'sphinxcontrib.napoleon', 'sphinx.ext.autodoc',
-              'sphinxemoji.sphinxemoji']
+extensions = [
+    "sphinx.ext.viewcode",
+    "sphinx.ext.todo",
+    "sphinxcontrib.napoleon",
+    "sphinx.ext.autodoc",
+    "sphinxemoji.sphinxemoji",
+]
 numfig = True
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -75,13 +83,13 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = None
-pygments_style = 'monokai'
+pygments_style = "monokai"
 # pygments_style = 'default'
-sphinxemoji_style = 'twemoji'
+sphinxemoji_style = "twemoji"
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = False
[…]
