### 2020-03-27 #2083 @tristan-salles
Hi @sjpfenninger sorry forgot to remove the setup.py and setup.cfg in the src/ folder. Just did it and tested that the installation via setup works as expected.
