+++ src/README.md
@@ -1,119 +1,0 @@
-# RADWave - _Wave analysis from Altimeter data_
-
-
-[![Docker Cloud Automated build](https://img.shields.io/docker/automated/pyreefmodel/radwave)](https://hub.docker.com/r/pyreefmodel/radwave)
-[![PyPI](https://img.shields.io/pypi/v/RADWave)](https://pypi.org/project/RADWave/) [![Documentation Status](https://readthedocs.org/projects/radwave/badge/?version=latest)](https://radwave.readthedocs.io/en/latest/?badge=latest)
-
-[![DOI](https://zenodo.org/badge/235496363.svg)](https://zenodo.org/badge/latestdoi/235496363)
-
-**RADWave** documentation is found at [**radwave.readthedocs.io**](https://radwave.readthedocs.io/)
-
-**RADWave** is a python package built to characterise *wave conditions* based on altimeter data.
-
-> **Satellite radar altimeters** can be used to determine significant wave height and wind speed [Young et al., 2011]. Since the first launch of the **GEOSAT** (GEOdetic SATellite) altimeter in 1985, there has been almost continuous data collection. From these measurements one can calculate wave conditions namely wave height, period and power [Young and Donelan, 2018]. Analysis of this long-term, high resolution spatio-temporal record brings new insights into inter-annual, seasonal and decadal variations of regional wave climates.
-
-![altimeter](https://github.com/pyReef-model/RADWave/blob/master/src/RADWave/Notebooks/images/img2.jpg?raw=true)
-
-**RADWave** is Python package that provides a mechanism to access altimeter datasets through web-enabled data services (THREDDS). The package capabilities are illustrated based on the the Australian Ocean Data Network ([**AODN**](https://portal.aodn.org.au) database that spans from 1985-present and that has already been calibrated and validated by [Ribal and Young, 2019]. **RADWave** allows the user to query over a range of spatial and temporal scales altimeter parameters in specific geographical regions and subsequently calculates significant wave heights, periods, group velocities, average wave energy densities and wave energy fluxes.  
-
-**RADWave** can be used to easily calculate past wave conditions and infers long term wave climate variability, providing new insights on wave modal conditions, seasonal changes, long-term trends and associated modulation by climate oscillations.
-
-Designed for researchers and industry partners focusing on offshore wave conditions globally, **RADWave** enhances the ease of access and analysis of altimeter data.
-
-## Installation
-
-### Dependencies
-
-You will need **Python 3.5+**.
-Also, the following packages are required:
-
- - [`numpy`](http://numpy.org)
- - [`scipy`](https://scipy.org)
- - [`pandas`](https://pandas.pydata.org/)
- - [`scikit-image`](https://scikit-image.org/)
- - [`seaborn`](https://seaborn.pydata.org)
- - [`geopy`](https://pypi.org/project/geopy/)
- - [`cartopy`](https://scitools.org.uk/cartopy/docs/latest/)
- - [`netCDF4`](https://pypi.org/project/netCDF4/)
- - [`shapely`](https://pypi.org/project/Shapely/)
- - [`pymannkendall`](https://pypi.org/project/pymannkendall/)
-
-### Installing using pip
-
-You can install `RADWave` using the
-[`pip package manager`](https://pypi.org/project/pip/) with your version of Python:
-
-```bash
-python3 -m pip install radwave
-```
-
-### Installing using Docker
-
-A more straightforward installation which does not depend on specific compilers relies on the [docker](http://www.docker.com) virtualisation system.
-
-To install the docker image and test it is working:
-
-```bash
-   docker pull pyreefmodel/radwave:latest
-   docker run --rm pyreefmodel/radwave:latest help
-```
-
[…]
