+++ setup.py
@@ -27,7 +27,7 @@ if __name__ == "__main__":
                     satellite altimeter data.",
         long_description=long_description,
         long_description_content_type="text/markdown",
-        packages=["src/RADWave"],
+        packages=["RADWave"],
         install_requires=[
             "pytest",
             "numpy>=1.15.0",
