### 2020-03-27 c755ebc
remove setup from root directory
