+++ tests/test_1_wave_analysis.py
@@ -1,106 +1,130 @@
 import pytest
 import RADWave
 
+
 def test_waveAnalysis_class_building():
 
-    wclass = RADWave.waveAnalysis(altimeterURL='tests/testURLs.txt',
-                                  bbox=[152.0, 153.0, -36.0, -35.0],
-                                  stime=[1998, 1, 1],
-                                  etime=[2008, 12, 31]
-                                 )
+    wclass = RADWave.waveAnalysis(
+        altimeterURL="tests/testURLs.txt",
+        bbox=[152.0, 153.0, -36.0, -35.0],
+        stime=[1998, 1, 1],
+        etime=[2008, 12, 31],
+    )
 
-    satName = ['JASON-2', 'ENVISAT', 'TOPEX']
-    assert wclass.satNb == 10, "test failed because satellites number equals " + str(wclass.satNb) + " instead of 10."
+    satName = ["JASON-2", "ENVISAT", "TOPEX"]
+    assert wclass.satNb == 10, (
+        "test failed because satellites number equals "
+        + str(wclass.satNb)
+        + " instead of 10."
+    )
     assert wclass.nameSat == satName, "test failed because satellites names missmatch "
+
 
 def test_altimeter_processing():
 
-    wclass = RADWave.waveAnalysis(altimeterURL='tests/testURLs.txt',
-                                  bbox=[152.0, 155.0, -36.0, -34.0],
-                                  stime=[1998, 1, 1],
-                                  etime=[2008, 12, 31]
-                                 )
+    wclass = RADWave.waveAnalysis(
+        altimeterURL="tests/testURLs.txt",
+        bbox=[152.0, 155.0, -36.0, -34.0],
+        stime=[1998, 1, 1],
+        etime=[2008, 12, 31],
+    )
 
-    wclass.processAltimeterData(altimeter_pick='all', saveCSV = 'altimeterData.csv')
-    assert pytest.approx(wclass.wh.mean(),rel=1e-3) == 2.386, "test failed because extracted mean wave height missmatch"
-    assert pytest.approx(wclass.ws.mean(),rel=1e-3) == 8.221, "test failed because extracted mean wind speed missmatch"
+    wclass.readAltimeterData(saveCSV="tests/altimeterData1.csv")
+    # wclass.processAltimeterData(altimeter_pick="all", saveCSV='tests/altimeterData1.csv')
+    assert (
+        pytest.approx(wclass.wh.mean(), rel=1e-3) == 2.386
+    ), "test failed because extracted mean wave height missmatch"
+    assert (
+        pytest.approx(wclass.ws.mean(), rel=1e-3) == 8.221
+    ), "test failed because extracted mean wind speed missmatch"
+
 
 def test_timeseries_generation():
 
-    wclass = RADWave.waveAnalysis(altimeterURL='tests/testURLs.txt',
-                                  bbox=[152.0, 155.0, -36.0, -34.0],
-                                  stime=[1998, 1, 1],
-                                  etime=[2008, 12, 31]
-                                 )
+    wclass = RADWave.waveAnalysis(
+        altimeterURL="tests/testURLs.txt",
+        bbox=[152.0, 155.0, -36.0, -34.0],
+        stime=[1998, 1, 1],
+        etime=[2008, 12, 31],
+    )
 
-    wclass.readAltimeterData(saveCSV = 'altimeterData.csv')
+    wclass.readAltimeterData(saveCSV="tests/altimeterData1.csv")
     ts = wclass.generateTimeSeries(days=30)
 
-    assert pytest.approx(ts['power'].mean(),rel=1e-3) == 87.752, "test failed"
[…]
