+++ tests/test_2_cyclone_tracks.py
@@ -1,62 +1,83 @@
 import pytest
 import RADWave
 
+
 def test_trackCyclone_class_building():
 
-    cyc = RADWave.waveAnalysis(altimeterURL='tests/IMOS_YASI.txt',
-                               bbox=[170, 175, -17, -12],
-                               stime=[2011,1,27],
-                               etime=[2011,2,4],
-                               cycloneCSV='tests/2010-YASI.csv'
-                              )
+    cyc = RADWave.waveAnalysis(
+        altimeterURL="tests/IMOS_YASI.txt",
+        bbox=[170, 175, -17, -12],
+        stime=[2011, 1, 27],
+        etime=[2011, 2, 4],
+        cycloneCSV="tests/2010-YASI.csv",
+    )
 
-    assert round(cyc.cyclone['pressure'].mean(),3) == 977.214, "test failed "
-    assert cyc.cyclone['name'][0] == 'YASI', "test failed "
+    assert round(cyc.cyclone["pressure"].mean(), 3) == 977.214, "test failed "
+    assert cyc.cyclone["name"][0] == "YASI", "test failed "
 
-    satName = ['JASON-2', 'CRYOSAT-2', 'ENVISAT']
-    assert cyc.satNb == 10, "test failed because satellites number equals " + str(wclass.satNb) + " instead of 10."
+    satName = ["JASON-2", "CRYOSAT-2", "ENVISAT"]
+    assert cyc.satNb == 10, (
+        "test failed because satellites number equals "
+        + str(wclass.satNb)
+        + " instead of 10."
+    )
     assert cyc.nameSat == satName, "test failed because satellites names missmatch "
+
 
 def test_cycloneAltimeter_processing():
 
-    cyc = RADWave.waveAnalysis(altimeterURL='tests/IMOS_YASI.txt',
-                               bbox=[170, 175, -17, -12],
-                               stime=[2011,1,27],
-                               etime=[2011,2,4],
-                               cycloneCSV='tests/2010-YASI.csv'
-                              )
+    cyc = RADWave.waveAnalysis(
+        altimeterURL="tests/IMOS_YASI.txt",
+        bbox=[170, 175, -17, -12],
+        stime=[2011, 1, 27],
+        etime=[2011, 2, 4],
+        cycloneCSV="tests/2010-YASI.csv",
+    )
 
-    cyc.processAltimeterData(altimeter_pick='all', saveCSV = 'altimeterData.csv')
-    assert pytest.approx(cyc.wh.mean(),rel=1e-3) == 2.292, "test failed because extracted mean wave height missmatch"
-    assert pytest.approx(cyc.ws.mean(),rel=1e-3) == 8.002, "test failed because extracted mean wind speed missmatch"
+    cyc.readAltimeterData(saveCSV="tests/altimeterData2.csv")
+    # cyc.processAltimeterData(altimeter_pick="all", saveCSV="tests/altimeterData2.csv")
+    assert (
+        pytest.approx(cyc.wh.mean(), rel=1e-3) == 2.292
+    ), "test failed because extracted mean wave height missmatch"
+    assert (
+        pytest.approx(cyc.ws.mean(), rel=1e-3) == 8.002
+    ), "test failed because extracted mean wind speed missmatch"
+
 
 def test_cycloneTimeseries_generation():
 
-    cyc = RADWave.waveAnalysis(altimeterURL='tests/IMOS_YASI.txt',
-                               bbox=[170, 175, -17, -12],
-                               stime=[2011,1,27],
-                               etime=[2011,2,4],
[…]
