+++ tests/test_1_wave_analysis.py
@@ -29,8 +29,10 @@ def test_altimeter_processing():
         etime=[2008, 12, 31],
     )
 
-    wclass.readAltimeterData(saveCSV="tests/altimeterData1.csv")
-    # wclass.processAltimeterData(altimeter_pick="all", saveCSV='tests/altimeterData1.csv')
+    # wclass.readAltimeterData(saveCSV="tests/altimeterData1.csv")
+    wclass.processAltimeterData(
+        altimeter_pick="all", saveCSV="tests/altimeterData1.csv"
+    )
     assert (
         pytest.approx(wclass.wh.mean(), rel=1e-3) == 2.386
     ), "test failed because extracted mean wave height missmatch"
