+++ tests/test_2_cyclone_tracks.py
@@ -34,8 +34,8 @@ def test_cycloneAltimeter_processing():
         cycloneCSV="tests/2010-YASI.csv",
     )
 
-    cyc.readAltimeterData(saveCSV="tests/altimeterData2.csv")
-    # cyc.processAltimeterData(altimeter_pick="all", saveCSV="tests/altimeterData2.csv")
+    # cyc.readAltimeterData(saveCSV="tests/altimeterData2.csv")
+    cyc.processAltimeterData(altimeter_pick="all", saveCSV="tests/altimeterData2.csv")
     assert (
         pytest.approx(cyc.wh.mean(), rel=1e-3) == 2.292
     ), "test failed because extracted mean wave height missmatch"
