+++ docs/conf.py
@@ -17,7 +17,6 @@ import sys
 
 sys.path.insert(0, os.path.abspath("."))
 sys.path.insert(0, os.path.abspath("../"))
-# sys.path.insert(0, os.path.abspath("../src"))
 
 import sphinx_rtd_theme
 
