+++ requirements.txt
@@ -2,7 +2,7 @@ numpy>=1.16.1
 pytest
 six>=1.11.0
 setuptools>=38.4.0
-pandas>=0.25
+pandas==0.25.0
 seaborn>=0.9
 matplotlib>=3.0
 geopy>=1.20
@@ -12,3 +12,4 @@ netCDF4>=1.5.1
 shapely>=1.6.4
 scikit-image>=0.15
 pymannkendall>=0
+cftime==1.0.3.4
