### 2020-03-29 #2083 @tristan-salles
@sjpfenninger, you were right, I did forget to remove the setup files in the src. I have made the changes to ensure that there were no more duplications.
