+++ joss/paper.md
@@ -32,9 +32,9 @@ Additionally, the high spatial resolution provided by altimeter data is particul
 
 # RADWave
 
-![An example of wave analysis perform with RADWave and its plotting capabilities.\label{fig:example}](fig1.jpg)
+![An example of the wave analysis and plotting capabilities of RADWave.\label{fig:example}](fig1.jpg)
 
-**RADWave** is Python package that provides a mechanism to access altimeter datasets through web-enabled data services (THREDDS). The package capabilities are illustrated based on the the Australian Ocean Data Network global database [@Ribal:2019]. **RADWave** allows  the user to query over a range of spatial and temporal scales altimeter parameters in specific geographical regions and subsequently calculates significant wave heights, periods, group velocities, average wave energy densities and wave energy fluxes.
+**RADWave** is Python package that provides a mechanism to access altimeter datasets through web-enabled data services (THREDDS). The package capabilities are illustrated in Figure 1 based on the the Australian Ocean Data Network global database [@Ribal:2019]. **RADWave** allows  the user to query over a range of spatial and temporal scales altimeter parameters in specific geographical regions and subsequently calculates significant wave heights, periods, group velocities, average wave energy densities and wave energy fluxes.
 
 While domain experts are able to assess satellite radar altimetry datasets, they currently have to download the data manually and write custom analysis and visualization functions using, e.g., numpy, pandas, and matplotlib to extract wave characteristics. **RADWave** combines this all into a simple object-oriented framework.
 
@@ -42,10 +42,10 @@ While domain experts are able to assess satellite radar altimetry datasets, they
 
 Along with the documentation a series of Jupyter Notebooks are presented to illustrate the package capability. **RADWave** enhances the ease of access and analysis of altimeter data and is designed for researchers and industry focusing on offshore wave conditions globally. To the authors' knowledge, no open-source code currently exists that provides such capability.
 
-**RADWave** is currently been applied for the characterisation of wave climate and extreme events offshore the Great Barrier Reef and to determine the magnitude and resulting impacts of extreme wave events offshore One Tree Island, Great Barrier Reef.
+**RADWave** is currently being applied to the characterisation of wave climate and extreme events offshore the Great Barrier Reef and to determine the magnitude and resulting impacts of extreme wave events offshore One Tree Island, Great Barrier Reef.
 
 # Acknowledgements
 
-The authors acknowledge the Integrated Marine Observing System (IMOS) a national collaborative research infrastructure supported by Australian Government and the Australian Ocean Data Network portal (**AODN**) for encouraging and developing the culture of data sharing across the marine science community of Australia.
+The authors acknowledge the Integrated Marine Observing System (IMOS), a national collaborative research infrastructure supported by Australian Government and the Australian Ocean Data Network portal (**AODN**), for encouraging and developing the culture of data sharing across the marine science community of Australia.
 
 # References
