### 2020-03-30 #2083 @sjpfenninger
@tristan-salles I suggested some minor changes to the paper here: https://github.com/pyReef-model/RADWave/pull/2
