### 2020-03-30 #2083 @sjpfenninger
@tristan-salles Your only tagged release on GitHub is v0.0.2 but PyPI has v0.1.1. Can you make sure you have a tagged release for the most recent version, and report back the correct version number here? Please then make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the Zenodo/figshare archive, please make sure that:

* The title of the archive is the same as the JOSS paper title
* That the authors of the archive are the same as the JOSS paper authors

Once that's done, we can move forward with accepting this!
