+++ Docker/Dockerfile
@@ -17,7 +17,7 @@ RUN apt-get update -qq && \
 
 ### RADWave - Notebooks
 RUN python3 -m pip install --no-cache-dir \
-    radwave==0.1.1
+    radwave==0.1.2
 
 RUN git clone https://github.com/pyReef-model/RADWave
 
