### 2020-03-31 #2083 @tristan-salles
Thanks @sjpfenninger! I have merged the PR and update the version number to 1.0.0 as well as attached it to a new Zenodo archive

[![DOI](https://zenodo.org/badge/235496363.svg)](https://zenodo.org/badge/latestdoi/235496363)

+ the title of the archive is the same as the JOSS paper title
+ the authors of the archive are the same as the JOOS paper authors
